Move much of the v1alpha1 utilities under test/v1alpha1
I generally consolidated `foo` related things under `test/v1alpha1/foo.go`.
