+++ test/conformance/api/cmd_args_test.go
@@ -24,6 +24,7 @@ import (
 
 	"github.com/knative/serving/pkg/apis/serving/v1alpha1"
 	"github.com/knative/serving/test"
+	v1a1test "github.com/knative/serving/test/v1alpha1"
 )
 
 func TestCmdArgsService(t *testing.T) {
@@ -42,7 +43,7 @@ func TestCmdArgsService(t *testing.T) {
 	const text = "THIS IS THE CMD AND ARGS TEST"
 
 	// Setup initial Service
-	_, err := test.CreateRunLatestServiceReady(t, clients, &names, &test.Options{},
+	_, err := v1a1test.CreateRunLatestServiceReady(t, clients, &names, &v1a1test.Options{},
 		func(svc *v1alpha1.Service) {
 			c := &svc.Spec.Template.Spec.Containers[0]
 			c.Image = names.Image
