+++ test/conformance/api/configuration_test.go
@@ -24,6 +24,7 @@ import (
 
 	"github.com/knative/serving/pkg/apis/serving/v1alpha1"
 	"github.com/knative/serving/test"
+	v1a1test "github.com/knative/serving/test/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -40,7 +41,7 @@ func TestUpdateConfigurationMetadata(t *testing.T) {
 	test.CleanupOnInterrupt(func() { test.TearDown(clients, names) })
 
 	t.Logf("Creating new configuration %s", names.Config)
-	if _, err := test.CreateConfiguration(t, clients, names, &test.Options{}); err != nil {
+	if _, err := v1a1test.CreateConfiguration(t, clients, names, &v1a1test.Options{}); err != nil {
 		t.Fatalf("Failed to create configuration %s", names.Config)
 	}
 
@@ -84,7 +85,7 @@ func TestUpdateConfigurationMetadata(t *testing.T) {
 	}
 
 	t.Logf("Validating labels were not propagated to Revision %s", names.Revision)
-	err = test.CheckRevisionState(clients.ServingClient, names.Revision, func(r *v1alpha1.Revision) (bool, error) {
+	err = v1a1test.CheckRevisionState(clients.ServingClient, names.Revision, func(r *v1alpha1.Revision) (bool, error) {
 		// Labels we placed on Configuration should _not_ appear on Revision.
 		return checkNoKeysPresent(newLabels, r.Labels, t), nil
 	})
@@ -122,7 +123,7 @@ func TestUpdateConfigurationMetadata(t *testing.T) {
 	}
 
 	t.Logf("Validating annotations were not propagated to Revision %s", names.Revision)
-	err = test.CheckRevisionState(clients.ServingClient, names.Revision, func(r *v1alpha1.Revision) (bool, error) {
+	err = v1a1test.CheckRevisionState(clients.ServingClient, names.Revision, func(r *v1alpha1.Revision) (bool, error) {
 		// Annotations we placed on Configuration should _not_ appear on Revision.
 		return checkNoKeysPresent(newAnnotations, r.Annotations, t), nil
 	})
@@ -141,7 +142,7 @@ func fetchConfiguration(name string, clients *test.Clients, t *testing.T) *v1alp
 
 func waitForConfigurationLatestCreatedRevision(clients *test.Clients, names test.ResourceNames) (string, error) {
 	var revisionName string
-	err := test.WaitForConfigurationState(clients.ServingClient, names.Config, func(c *v1alpha1.Configuration) (bool, error) {
+	err := v1a1test.WaitForConfigurationState(clients.ServingClient, names.Config, func(c *v1alpha1.Configuration) (bool, error) {
 		if c.Status.LatestCreatedRevisionName != names.Revision {
 			revisionName = c.Status.LatestCreatedRevisionName
 			return true, nil
@@ -152,13 +153,13 @@ func waitForConfigurationLatestCreatedRevision(clients *test.Clients, names test
 }
 
 func waitForConfigurationLabelsUpdate(clients *test.Clients, names test.ResourceNames, labels map[string]string) error {
-	return test.WaitForConfigurationState(clients.ServingClient, names.Config, func(c *v1alpha1.Configuration) (bool, error) {
+	return v1a1test.WaitForConfigurationState(clients.ServingClient, names.Config, func(c *v1alpha1.Configuration) (bool, error) {
 		return reflect.DeepEqual(c.Labels, labels) && c.Generation == c.Status.ObservedGeneration, nil
 	}, "ConfigurationMetadataUpdatedWithLabels")
 }
 
 func waitForConfigurationAnnotationsUpdate(clients *test.Clients, names test.ResourceNames, annotations map[string]string) error {
-	return test.WaitForConfigurationState(clients.ServingClient, names.Config, func(c *v1alpha1.Configuration) (bool, error) {
+	return v1a1test.WaitForConfigurationState(clients.ServingClient, names.Config, func(c *v1alpha1.Configuration) (bool, error) {
 		return reflect.DeepEqual(c.Annotations, annotations) && c.Generation == c.Status.ObservedGeneration, nil
 	}, "ConfigurationMetadataUpdatedWithAnnotations")
 }
