+++ test/conformance/runtime/user_test.go
@@ -22,6 +22,7 @@ import (
 	"testing"
 
 	"github.com/knative/serving/test"
+	v1a1test "github.com/knative/serving/test/v1alpha1"
 
 	corev1 "k8s.io/api/core/v1"
 )
@@ -42,7 +43,7 @@ func TestMustRunAsUser(t *testing.T) {
 		RunAsUser: &runAsUser,
 	}
 
-	_, ri, err := fetchRuntimeInfo(t, clients, &test.Options{SecurityContext: securityContext})
+	_, ri, err := fetchRuntimeInfo(t, clients, &v1a1test.Options{SecurityContext: securityContext})
 	if err != nil {
 		t.Fatalf("Error fetching runtime info: %v", err)
 	}
@@ -72,7 +73,7 @@ func TestMustRunAsUser(t *testing.T) {
 func TestShouldRunAsUserContainerDefault(t *testing.T) {
 	t.Parallel()
 	clients := test.Setup(t)
-	_, ri, err := fetchRuntimeInfoUnprivileged(t, clients, &test.Options{})
+	_, ri, err := fetchRuntimeInfoUnprivileged(t, clients, &v1a1test.Options{})
 
 	if err != nil {
 		t.Fatalf("Error fetching runtime info: %v", err)
