+++ test/crd.go
@@ -22,23 +22,10 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/knative/pkg/ptr"
-	ptest "github.com/knative/pkg/test"
 	"github.com/knative/pkg/test/helpers"
-	"github.com/knative/serving/pkg/apis/serving/v1alpha1"
-	"github.com/knative/serving/pkg/apis/serving/v1beta1"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	v1alpha1testing "github.com/knative/serving/pkg/testing/v1alpha1"
 )
 
 const (
-	// Default for user containers in e2e tests. This value is lower than the general
-	// Knative's default so as to run more effectively in CI with limited resources.
-	defaultRequestCPU = "100m"
-
 	testNamePrefix = "Test"
 )
 
@@ -51,166 +38,6 @@ type ResourceNames struct {
 	TrafficTarget string
 	Domain        string
 	Image         string
-}
-
-// ResourceObjects holds types of the resource objects.
-type ResourceObjects struct {
-	Route    *v1alpha1.Route
-	Config   *v1alpha1.Configuration
-	Service  *v1alpha1.Service
-	Revision *v1alpha1.Revision
-}
-
-// Route returns a Route object in namespace using the route and configuration
-// names in names.
-func Route(names ResourceNames, fopt ...v1alpha1testing.RouteOption) *v1alpha1.Route {
-	route := &v1alpha1.Route{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: names.Route,
-		},
-		Spec: v1alpha1.RouteSpec{
-			Traffic: []v1alpha1.TrafficTarget{{
-				TrafficTarget: v1beta1.TrafficTarget{
-					Tag:               names.TrafficTarget,
-					ConfigurationName: names.Config,
-					Percent:           100,
-				},
-			}},
-		},
-	}
-
-	for _, opt := range fopt {
-		opt(route)
-	}
-
-	return route
-}
-
-// ConfigurationSpec returns the spec of a configuration to be used throughout different
-// CRD helpers.
-func ConfigurationSpec(imagePath string, options *Options) *v1alpha1.ConfigurationSpec {
-	if options.ContainerResources.Limits == nil && options.ContainerResources.Requests == nil {
-		options.ContainerResources = corev1.ResourceRequirements{
-			Requests: corev1.ResourceList{
-				corev1.ResourceCPU: resource.MustParse(defaultRequestCPU),
-			},
-		}
-	}
-
-	spec := &v1alpha1.ConfigurationSpec{
-		Template: &v1alpha1.RevisionTemplateSpec{
-			Spec: v1alpha1.RevisionSpec{
-				RevisionSpec: v1beta1.RevisionSpec{
-					PodSpec: corev1.PodSpec{
-						Containers: []corev1.Container{{
-							Image:           imagePath,
-							Resources:       options.ContainerResources,
-							ReadinessProbe:  options.ReadinessProbe,
-							Ports:           options.ContainerPorts,
-							SecurityContext: options.SecurityContext,
-						}},
-					},
-					ContainerConcurrency: v1beta1.RevisionContainerConcurrencyType(options.ContainerConcurrency),
-				},
-			},
-		},
-	}
-
-	if options.RevisionTimeoutSeconds > 0 {
-		spec.GetTemplate().Spec.TimeoutSeconds = ptr.Int64(options.RevisionTimeoutSeconds)
-	}
-
-	if options.EnvVars != nil {
-		spec.GetTemplate().Spec.GetContainer().Env = options.EnvVars
-	}
-
-	if len(options.RevisionTemplateAnnotations) != 0 {
-		spec.Template.ObjectMeta = metav1.ObjectMeta{
-			Annotations: options.RevisionTemplateAnnotations,
-		}
-	}
-
-	return spec
-}
-
-// LegacyConfigurationSpec returns the spec of a configuration to be used throughout different
-// CRD helpers.
-func LegacyConfigurationSpec(imagePath string, options *Options) *v1alpha1.ConfigurationSpec {
-	if options.ContainerResources.Limits == nil && options.ContainerResources.Requests == nil {
-		options.ContainerResources = corev1.ResourceRequirements{
-			Requests: corev1.ResourceList{
-				corev1.ResourceCPU: resource.MustParse(defaultRequestCPU),
-			},
-		}
-	}
-
-	spec := &v1alpha1.ConfigurationSpec{
-		DeprecatedRevisionTemplate: &v1alpha1.RevisionTemplateSpec{
-			Spec: v1alpha1.RevisionSpec{
-				DeprecatedContainer: &corev1.Container{
-					Image:           imagePath,
-					Resources:       options.ContainerResources,
-					ReadinessProbe:  options.ReadinessProbe,
-					Ports:           options.ContainerPorts,
-					SecurityContext: options.SecurityContext,
-				},
-				RevisionSpec: v1beta1.RevisionSpec{
-					ContainerConcurrency: v1beta1.RevisionContainerConcurrencyType(options.ContainerConcurrency),
-				},
-			},
-		},
-	}
-
-	if options.RevisionTimeoutSeconds > 0 {
-		spec.GetTemplate().Spec.TimeoutSeconds = ptr.Int64(options.RevisionTimeoutSeconds)
-	}
-
-	if options.EnvVars != nil {
-		spec.GetTemplate().Spec.GetContainer().Env = options.EnvVars
-	}
-
-	return spec
-}
-
-// Configuration returns a Configuration object in namespace with the name names.Config
-// that uses the image specified by names.Image
-func Configuration(names ResourceNames, options *Options, fopt ...v1alpha1testing.ConfigOption) *v1alpha1.Configuration {
-	config := &v1alpha1.Configuration{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: names.Config,
-		},
-		Spec: *ConfigurationSpec(ptest.ImagePath(names.Image), options),
-	}
-	if options.ContainerPorts != nil && len(options.ContainerPorts) > 0 {
-		config.Spec.GetTemplate().Spec.GetContainer().Ports = options.ContainerPorts
-	}
-
-	for _, opt := range fopt {
-		opt(config)
-	}
-
-	return config
-}
-
-// LatestService returns a Service object in namespace with the name names.Service
-// that uses the image specified by names.Image.
-func LatestService(names ResourceNames, options *Options, fopt ...v1alpha1testing.ServiceOption) *v1alpha1.Service {
-	a := append([]v1alpha1testing.ServiceOption{
-		v1alpha1testing.WithInlineConfigSpec(*ConfigurationSpec(ptest.ImagePath(names.Image), options)),
-	}, fopt...)
-	return v1alpha1testing.ServiceWithoutNamespace(names.Service, a...)
-}
-
-// LatestServiceLegacy returns a DeprecatedRunLatest Service object in namespace with the name names.Service
-// that uses the image specified by names.Image.
-func LatestServiceLegacy(names ResourceNames, options *Options, fopt ...v1alpha1testing.ServiceOption) *v1alpha1.Service {
-	a := append([]v1alpha1testing.ServiceOption{
-		v1alpha1testing.WithRunLatestConfigSpec(*LegacyConfigurationSpec(ptest.ImagePath(names.Image), options)),
-	}, fopt...)
-	svc := v1alpha1testing.ServiceWithoutNamespace(names.Service, a...)
-	// Clear the name, which is put there by defaulting.
-	svc.Spec.DeprecatedRunLatest.Configuration.GetTemplate().Spec.GetContainer().Name = ""
-	return svc
 }
 
 // AppendRandomString will generate a random string that begins with prefix. This is useful
