+++ test/e2e/e2e.go
@@ -13,6 +13,7 @@ import (
 	pkgTest "github.com/knative/pkg/test"
 	"github.com/knative/serving/pkg/autoscaler"
 	"github.com/knative/serving/test"
+	v1a1test "github.com/knative/serving/test/v1alpha1"
 	perrors "github.com/pkg/errors"
 )
 
@@ -50,7 +51,7 @@ func SetupWithNamespace(t *testing.T, namespace string) *test.Clients {
 
 // CreateRouteAndConfig will create Route and Config objects using clients.
 // The Config object will serve requests to a container started from the image at imagePath.
-func CreateRouteAndConfig(t *testing.T, clients *test.Clients, image string, options *test.Options) (test.ResourceNames, error) {
+func CreateRouteAndConfig(t *testing.T, clients *test.Clients, image string, options *v1a1test.Options) (test.ResourceNames, error) {
 	svcName := test.ObjectNameForTest(t)
 	names := test.ResourceNames{
 		Config: svcName,
@@ -58,10 +59,10 @@ func CreateRouteAndConfig(t *testing.T, clients *test.Clients, image string, opt
 		Image:  image,
 	}
 
-	if _, err := test.CreateConfiguration(t, clients, names, options); err != nil {
+	if _, err := v1a1test.CreateConfiguration(t, clients, names, options); err != nil {
 		return test.ResourceNames{}, err
 	}
-	_, err := test.CreateRoute(t, clients, names)
+	_, err := v1a1test.CreateRoute(t, clients, names)
 	return names, err
 }
