--- test/states.go
@@ -1,101 +1,0 @@
-/*
-Copyright 2018 The Knative Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package test
-
-import (
-	"fmt"
-
-	"github.com/davecgh/go-spew/spew"
-	corev1 "k8s.io/api/core/v1"
-
-	"github.com/knative/serving/pkg/apis/serving"
-	"github.com/knative/serving/pkg/apis/serving/v1alpha1"
-)
-
-// states contains functions for asserting against the state of Knative Serving
-// crds to see if they have achieved the states specified in the spec
-// (https://github.com/knative/serving/blob/master/docs/spec/spec.md).
-
-// AllRouteTrafficAtRevision will check the revision that route r is routing
-// traffic to and return true if 100% of the traffic is routing to revisionName.
-func AllRouteTrafficAtRevision(names ResourceNames) func(r *v1alpha1.Route) (bool, error) {
-	return func(r *v1alpha1.Route) (bool, error) {
-		for _, tt := range r.Status.Traffic {
-			if tt.Percent == 100 {
-				if tt.RevisionName != names.Revision {
-					return true, fmt.Errorf("expected traffic revision name to be %s but actually is %s: %s", names.Revision, tt.RevisionName, spew.Sprint(r))
-				}
-
-				if tt.Tag != names.TrafficTarget {
-					return true, fmt.Errorf("expected traffic target name to be %s but actually is %s: %s", names.TrafficTarget, tt.Tag, spew.Sprint(r))
-				}
-
-				return true, nil
-			}
-		}
-		return false, nil
-	}
-}
-
-// IsRevisionReady will check the status conditions of the revision and return true if the revision is
-// ready to serve traffic. It will return false if the status indicates a state other than deploying
-// or being ready. It will also return false if the type of the condition is unexpected.
-func IsRevisionReady(r *v1alpha1.Revision) (bool, error) {
-	return r.Generation == r.Status.ObservedGeneration && r.Status.IsReady(), nil
-}
-
-// IsServiceReady will check the status conditions of the service and return true if the service is
-// ready. This means that its configurations and routes have all reported ready.
-func IsServiceReady(s *v1alpha1.Service) (bool, error) {
-	return s.Generation == s.Status.ObservedGeneration && s.Status.IsReady(), nil
-}
-
-// IsRouteReady will check the status conditions of the route and return true if the route is
-// ready.
-func IsRouteReady(r *v1alpha1.Route) (bool, error) {
-	return r.Generation == r.Status.ObservedGeneration && r.Status.IsReady(), nil
-}
-
-// ConfigurationHasCreatedRevision returns whether the Configuration has created a Revision.
-func ConfigurationHasCreatedRevision(c *v1alpha1.Configuration) (bool, error) {
-	return c.Status.LatestCreatedRevisionName != "", nil
-}
-
-// IsConfigRevisionCreationFailed will check the status conditions of the
-// configuration and return true if the configuration's revision failed to
-// create.
-func IsConfigRevisionCreationFailed(c *v1alpha1.Configuration) (bool, error) {
-	if cond := c.Status.GetCondition(v1alpha1.ConfigurationConditionReady); cond != nil {
-		return cond.Status == corev1.ConditionFalse && cond.Reason == "RevisionFailed", nil
-	}
-	return false, nil
-}
-
-// IsRevisionAtExpectedGeneration returns a function that will check if the annotations
-// on the revision include an annotation for the generation and that the annotation is
-// set to the expected value.
-func IsRevisionAtExpectedGeneration(expectedGeneration string) func(r *v1alpha1.Revision) (bool, error) {
-	return func(r *v1alpha1.Revision) (bool, error) {
-		if a, ok := r.Labels[serving.ConfigurationGenerationLabelKey]; ok {
-			if a != expectedGeneration {
-				return true, fmt.Errorf("expected Revision %s to be labeled with generation %s but was %s instead", r.Name, expectedGeneration, a)
-			}
-			return true, nil
-		}
-		return true, fmt.Errorf("expected Revision %s to be labeled with generation %s but there was no label", r.Name, expectedGeneration)
-	}
-}
